time util generalized
