+++ Core/Inc/top_board/Util/tim_util.h
@@ -9,6 +9,7 @@
 #define UTILS_TIM_UTIL_H_
 
 #include "stm32f7xx_hal.h"
+#include "tim_util_functions.h"
 
 // Timers (clocks still need to be defined: wheel_control, KickChip)
 extern TIM_HandleTypeDef htim1;     // encoder RF       16 bit
@@ -25,13 +26,6 @@ extern TIM_HandleTypeDef htim11;    // KickChipping
 extern TIM_HandleTypeDef htim12;    // PWM (CH1: Dribbler)
 extern TIM_HandleTypeDef htim13;    // internal timer (not used ATM)
 extern TIM_HandleTypeDef htim14;    // internal timer (not used ATM)
-
-
-// structs
-typedef struct PWM_struct{
-    TIM_HandleTypeDef* TIM;
-    uint16_t Channel;
-} PWM_struct;
 
 // Timer Helpers
 // Encoders
@@ -53,24 +47,5 @@ extern PWM_struct PWM_LF;
 extern PWM_struct PWM_LB;
 
 extern PWM_struct PWM_Buzzer;
-
-
-inline void start_PWM(PWM_struct PWM){
-    HAL_TIM_Base_Start(PWM.TIM);
-    HAL_TIM_PWM_Start(PWM.TIM, PWM.Channel);
-}
-
-inline void stop_PWM(PWM_struct PWM){
-    HAL_TIM_Base_Stop(PWM.TIM);
-    HAL_TIM_PWM_Stop(PWM.TIM, PWM.Channel);
-}
-
-inline void set_PWM(PWM_struct PWM, uint32_t Value){
-    __HAL_TIM_SET_COMPARE(PWM.TIM , PWM.Channel, Value);
-}
-
-inline uint32_t get_PWM(PWM_struct PWM){
-    return __HAL_TIM_GET_COMPARE(PWM.TIM , PWM.Channel);
-}
 
 #endif //UTILS_TIM_UTIL_H_
