+++ Core/Src/dribbler_board/Util/tim_util.c
@@ -11,7 +11,7 @@
 PWM_struct PWM_Dribbler_a	    = { PWM_DRIBBLER , ((uint16_t)TIM_CHANNEL_1)};
 PWM_struct PWM_Dribbler_b       = { PWM_DRIBBLER , ((uint16_t)TIM_CHANNEL_2)}; 
 
-void set_PWM(PWM_struct* PWM, float Value){
+void set_PWM_dribbler(PWM_struct* PWM, float Value){
     if (PWM->Channel == ((uint16_t)TIM_CHANNEL_1))
         PWM->TIM->Instance->CCR1 = (uint16_t)(Value * 1000);
     else
