+++ Core/Src/dribbler_board/peripherals/dribbler.c
@@ -45,12 +45,12 @@ void dribbler_DeInit(){
 void dribbler_SetSpeed(float speed){
 	if (BOARD_INITIALIZED){
 		if (speed > 0){
-			set_PWM(&PWM_Dribbler_a, 0);
-			set_PWM(&PWM_Dribbler_b, speed);
+			set_PWM_dribbler(&PWM_Dribbler_a, 0);
+			set_PWM_dribbler(&PWM_Dribbler_b, speed);
 		}
 		else{
-			set_PWM(&PWM_Dribbler_b, 0);
-			set_PWM(&PWM_Dribbler_a, -1*speed);
+			set_PWM_dribbler(&PWM_Dribbler_b, 0);
+			set_PWM_dribbler(&PWM_Dribbler_a, -1*speed);
 		}
 	}
 }
